# Upgrade text model to gpt-3.5-turbo (ChatGPT)