Generates text ~5x faster.

ChatGPT's underlying model, `gpt-3.5-turbo`, [became available in the API today](https://openai.com/blog/introducing-chatgpt-and-whisper-apis). It should cost 10x less per token. It's also much faster:

| | Before (`text-davinci-003`)  | After (`gpt-3.5-turbo`) |
| --- | ------------- | ------------- |
| Trial 1 | 3371.552 ms  | 622.586 ms  |
| Trial 2 | 2654.793 ms  | 688.590 ms  |
| Trial 3 | 2950.276 ms | 483.914 ms | 

I timed generations with the same prompt and tone using `console.time()` around the `reader.read()` loop in **index.tsx** [here](https://github.com/Nutlope/twitterbio/blob/main/pages/index.tsx#L60).

<img width="997" alt="image" src="https://user-images.githubusercontent.com/36117635/222270757-efd47bc9-e173-4b88-8fb9-2902b37c8141.png">

First PR in this repo; open to feedback on codestyle or PR etiquette!